moonbeam_client_evm_tracing
Module types

Runtime API allowing to debug/trace Ethereum

Modules

block

Types for tracing all Ethereum transactions of a block.

serialization

Provide serialization functions for various types and formats.

single

Types for the tracing of a single Ethereum transaction. Structure from “raw” debug_trace and a “call list” matching Blockscout formatter. This “call list” is also used to build the whole block tracing output.

Enums

CallResult

CallType

ContextType

CreateResult

CreateType

Constants

MANUAL_BLOCK_INITIALIZATION_RUNTIME_VERSION

Functions

convert_memory