moonbeam_client_evm_tracing::types
Module single

Types for the tracing of a single Ethereum transaction. Structure from “raw” debug_trace and a “call list” matching Blockscout formatter. This “call list” is also used to build the whole block tracing output.

Structs

Log

RawStepLog

TraceCallConfig

Enums

Call

TraceType

TransactionTrace

Single transaction trace.